// packages/core/src/rds/types/product.ts
import {
  ColumnType,
  Generated,
  Insertable,
  Selectable,
  Updateable,
} from "kysely";

export interface Database {
  product: ProductTable;
  product_image: ProductImageTable;
}

export interface ProductTable {
  // Columns that are generated by the database should be marked
  // using the `Generated` type. This way they are automatically
  // made optional in inserts and updates.
  product_id: Generated<string>;

  title: string;
  description: string;
  price: number;
}
export interface ProductImageTable {
  // Columns that are generated by the database should be marked
  // using the `Generated` type. This way they are automatically
  // made optional in inserts and updates.
  product_image_id: Generated<string>;
  product_id: string;
  key: string;
}

// You should not use the table schema interfaces directly. Instead, you should
// use the `Selectable`, `Insertable` and `Updateable` wrappers. These wrappers
// make sure that the correct types are used in each operation.
export type Product = Selectable<ProductTable>;
export type NewProduct = Insertable<ProductTable>;
export type UpdateProduct = Updateable<ProductTable>;

export type ProductImage = Selectable<ProductImageTable>;
export type NewProductImage = Insertable<ProductImageTable>;
export type UpdateProductImage = Updateable<ProductImageTable>;